backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yoojin Ko <yoo.ko@sky.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.6"
django-filter = "^23.2"
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
